feat(InputView): can handle users names input
- 레이싱 게임 플레이어가 입력하는 이름을 도메인이 맞게 다루기 위한 기능 구현
