+++ src/main/java/com/nextstep/racinggame/view/InputView.java
@@ -1,36 +1,18 @@
 package com.nextstep.racinggame.view;
 
 import com.nextstep.racinggame.domain.Name;
-import com.nextstep.racinggame.view.exceptions.InvalidUserInputException;
 
-import java.util.Arrays;
 import java.util.List;
-import java.util.stream.Collectors;
+import java.util.Scanner;
 
 public class InputView {
-    private static final String SEPARATOR = ",";
+    private static final Scanner SCANNER = new Scanner(System.in);
 
-    private final String userInput;
+    public static List<Name> getCarNames() {
+        String userInput = SCANNER.nextLine();
 
-    public InputView(final String userInput) {
-        validate(userInput);
+        NamesInputView namesInputView = new NamesInputView(userInput);
 
-        this.userInput = userInput;
-    }
-
-    public List<Name> parseToNames() {
-        return Arrays.stream(userInput.split(SEPARATOR))
-                .map(Name::new)
-                .collect(Collectors.toList());
-    }
-
-    private void validate(final String userInput) {
-        if (userInput == null) {
-            throw new InvalidUserInputException("Null로 객체를 생성할 수 없습니다.");
-        }
-
-        if (userInput.isEmpty()) {
-            throw new InvalidUserInputException("빈 문자열은 입력할 수 없습니다.");
-        }
+        return namesInputView.parseToNames();
     }
 }
